% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANPELA-package.R
\docType{package}
\name{ANPELA-package}
\alias{ANPELA}
\alias{ANPELA-package}
\title{ANPELA: Navigating the Data Processing for Single-cell Proteomics using ANPELA}
\description{
Single-cell proteomics (SCP) has emerged as a powerful technique that significantly advances our understanding of complex biological systems with new level of granularity. Because of the extreme difficulty in processing SCP data, ANPELA was developed for identifying the optimal workflow based on a well-designed assessment strategy. ANPELA is a user-centric and application-oriented tool which is capable of navigating the data processing for SCP. ANPELA 3.0 has significantly improved its practicality, focusing primarily on the following points: 1. multi-scenarios deployment (versatile choices meet diverse user needs); 2. data security (local execution ensures data confidentiality); 3. open source (modular codes facilitate readers’ free editing); and 4. user-friendly interface (the visual interface enhances user application). The local software and webserver of ANPELA are available at http://idrblab.org/anpela2024.
}
\author{
\strong{Maintainer}: Feng Zhu \email{zhufeng@zju.edu.cn}

Authors:
\itemize{
  \item Huaicheng Sun \email{sunhc@zju.edu.cn}
  \item Yuan Zhou
  \item Yuxuan Liu
}

}
\keyword{internal}
