% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Assess}
\alias{Assess}
\title{Assess}
\usage{
Assess(
  name,
  data = NULL,
  respath = NULL,
  studytype = c("CSI", "PTI"),
  clusteringM = "FlowSOM",
  ncluster = 8,
  Phenograph_k = 30,
  DEP = NULL,
  TIM = c("scorpius_distSpear", "scorpius_distPear", "scorpius_distEucl",
    "scorpius_distManh", "slingshot_FLOWMAP", "slingshot_tSNE", "prinCurves_tSNE",
    "slingshot_PCA", "slingshot_diffMaps", "prinCurves_diffMaps"),
  pathwayhierarchy = NULL,
  cores = floor(parallel::detectCores()/2),
  save_processed_res = c("no", "one_folder", "one_RData"),
  savepath = "./"
)
}
\arguments{
\item{name}{Character, the filename of all the files in the "assess_res" folder which will store the assessment results.}

\item{data}{Character, the resulting RData file of the function "Process", "FCprocess" or "MCprocess" when the "save_processed_res" parameter in these functions is set to "one_RData".}

\item{respath}{Character, the absolute path of the folder storing the resulting RData files of the function "Process", "FCprocess" or "MCprocess" when the "save_processed_res" parameter in these functions is set to "one_folder".}

\item{studytype}{Character, the type of your study, including "CSI (Cell Subpopulation Identification)" and "PTI (Pseudotime Trajectory Inference)".}

\item{clusteringM}{Character, the method of clustering the processed data prior to performance assessment, including "FlowSOM" and "PhenoGraph".
\if{html}{\out{<br>}}\strong{FlowSOM}: a widely used cluster clustering algorithm designed for high-dimensional cytometry data (the number of clusters needs to be specified). ANPELA uses the function “SOM” in R package "FlowSOM" to implement the algorithm.
\if{html}{\out{<br>}}\strong{PhenoGraph}: a well-designed clustering algorithm developed to define phenotypes in high-dimensional single-cell data (the number of clusters need not be specified). ANPELA uses the function “Rphenograph” in R package "cytofkit".}

\item{ncluster}{Integer, the number of clusters for meta clustering in FlowSOM.
\if{html}{\out{<br>}}Only needed when the argument of "clusteringM" is selected as "FlowSOM".}

\item{Phenograph_k}{Character, the number of nearest neighbours used in PhenoGraph clustering method.
\if{html}{\out{<br>}}Only needed when the argument of "clusteringM" is selected as "PhenoGraph".}

\item{DEP}{Character, the differentially expressed proteins used as the prior knowledge for the fourth criterion.}

\item{TIM}{Character, the method of trajectory inference for the processed data prior to performance assessment, consisted of trajectory reconstruction and data space representation, including "scorpius_distSpear", "scorpius_distPear", "scorpius_distManh","slingshot_tSNE", "prinCurves_tSNE", "slingshot_PCA", "slingshot_diffMaps" and "prinCurves_diffMaps".}

\item{pathwayhierarchy}{Character, the absolute filepath of the pathway hierarchy file.}

\item{cores}{Integer, the number of CPU cores to be employed for performing parallel computing.
\if{html}{\out{<br>}}To avoid memory explosion due to parallel computing, the default is the largest integers not greater than half of the number of CPU cores on the current host.}

\item{save_processed_res}{Character, the form of data processing output files. "no" denotes that the results would not be saved. "one_folder" denotes that successfully processed results will be saved as separate RData files in the "process_res" folder. "one_RData" denotes that all processed results will be saved as one RData file in the "process_res" folder.}

\item{savepath}{Character, the absolute path of the folder which will store files of the assessment and processed results.}
}
\description{
A short description...
}
\examples{
\donttest{
}
}
