% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CSIres.R
\name{Get_CSIres}
\alias{Get_CSIres}
\title{Get CSI (Cell Subpopulation Identification) Study Result(s)}
\usage{
Get_CSIres(
  respath,
  save_processed_res = "one_folder",
  workflow,
  savepath = "./ANPELA_res",
  marker_path,
  color = c("#F39B7FFF", "#3C5488FF", "#7E6148FF", "#B09C85FF", "#8491B4FF", "#4DBBD5FF",
    "#00A087FF", "#91D1C2FF", "#E64B35FF", "grey80", RColorBrewer::brewer.pal(12,
    "Set3")),
  plot = c(T, F)
)
}
\arguments{
\item{respath}{Character, the absolute path of the folder storing the resulting ‘info_saved.RData’ file and the ‘process_res’ folder of the function ‘Process’, ‘FCprocess’ or ‘MCprocess’ when the ‘save_processed_res’ parameter in these functions is set to ‘one_folder’.}

\item{save_processed_res}{Character, the format of the data processing output files. ‘one_folder’ denotes that successfully processed results will be saved as separate RData files in the ‘process_res’ folder. ‘one_RData’ denotes that all processed results will be saved as one RData file in the ‘process_res’ folder.}

\item{workflow}{Character, the combinations of data processing methods specified by users according to their research interests.
\if{html}{\out{<br>}}It is a vector includes one or more method combinations, typically in the format of ‘compensation method name_ transformation method name_ normalization method name_ signal clean method name ‘, for example: c(‘None_Biexponential Transformation_None_None’,’CytoSpill_FlowVS Transformation_None_FlowCut’).}

\item{savepath}{Character, the absolute path of the folder which will store the results.}

\item{marker_path}{Character, the absolute file path of the CSV or XLSX file containing the markers for cell type annotation, and detailed format requirements can be found on the website https://github.com/idrblab/ANPELA.}

\item{color}{Character, the vector of color specifications used to visually distinguish different cell annotation categories within the generated t-SNE plot.}

\item{plot}{Logical, the logical flag determining whether the t-SNE visualization plot based on the analysis results should be generated and saved as a PNG file.}
}
\value{
An RData file named "workflow_CSIres.RData", recording the CSI results. Optionally a PNG file in addition.
}
\description{
Get_CSIres() performs cell annotation and t-SNE dimensionality reduction on processed SCP data, optionally visualizing the resulting cell populations and saved as a PNG file.
}
\examples{
\donttest{
}
}
