% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_PTIres.R
\name{Get_PTIres}
\alias{Get_PTIres}
\title{Get PTI (Pseudotime Trajectory Inference) Study Result(s)}
\usage{
Get_PTIres(
  respath,
  save_processed_res = "one_folder",
  workflow,
  savepath = "./ANPELA_res",
  TIM = "scorpius_distSpear",
  clustering.var = NULL,
  plot = c(T, F)
)
}
\arguments{
\item{respath}{Character, the absolute path of the folder storing the resulting ‘info_saved.RData’ file and the ‘process_res’ folder of the function ‘Process’, ‘FCprocess’ or ‘MCprocess’ when the ‘save_processed_res’ parameter in these functions is set to ‘one_folder’.}

\item{save_processed_res}{Character, the format of the data processing output files. ‘one_folder’ denotes that successfully processed results will be saved as separate RData files in the ‘process_res’ folder. ‘one_RData’ denotes that all processed results will be saved as one RData file in the ‘process_res’ folder.}

\item{workflow}{Character, the combinations of data processing methods specified by users according to their research interests.
\if{html}{\out{<br>}}It is a vector includes one or more method combinations, typically in the format of ‘compensation method name_ transformation method name_ normalization method name_ signal clean method name ‘, for example: c(‘None_Biexponential Transformation_None_None’,’CytoSpill_FlowVS Transformation_None_FlowCut’).}

\item{savepath}{Character, the absolute path of the folder which will store the results.}

\item{TIM}{Character, the method of trajectory inference for the processed data prior to performance assessment, consisted of tra-jectory reconstruction and data space representation, including ‘scorpius_distSpear’, ‘scorpius_distPear’, ‘scorpius_distEucl’, ‘scorpius_distManh’, ‘slingshot_tSNE’, ‘slingshot_FLOWMAP’, ‘slingshot_PCA’, ‘slingshot_diffMaps’.}

\item{clustering.var}{Character, the vector naming channels to be used to calculate distances/differences between cells for clustering (if re-quested) and edge-drawing steps.}

\item{plot}{Logical, the logical flag determining whether the visualization plot representing the results of the selected trajectory inference method should be generated and saved as a PNG file.}
}
\value{
An RData file named "workflow_PTIres.RData", recording the PTI results. Optionally a PNG file in addition.
}
\description{
Get_PTIres() applies selected trajectory inference algorithms to processed SCP data, optionally visualizing the resulting cellular trajectories and saved as a PNG file.
}
\examples{
\donttest{
}
}
