% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getmarker.R
\name{Getmarker}
\alias{Getmarker}
\title{Get Marker}
\usage{
Getmarker(datapath)
}
\arguments{
\item{datapath}{Character, the absolute path of the folder storing the FCS raw data files and metadata file.}
}
\value{
A character including marker indexes for data processing and performance evaluation, with subsequent manual removal of non-protein columns as needed
}
\description{
getmarker enables the preparation of marker indexes for data processing and performance assessment, with subsequent manual removal of non-protein columns as needed.
}
\examples{
\donttest{
}
}
