% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_CSI.R
\name{MC_CSI}
\alias{MC_CSI}
\title{Mass Cytometry Cell Subpopulation Identification}
\usage{
MC_CSI(
  name = "result",
  datapath,
  mergeM = "Fixed",
  fixedNum = 200,
  compensationM = c("CATALYST", "CytoSpill", "spillR", "None"),
  transformationM = c("Arcsinh Transformation", "Asinh with Non-negative Value",
    "Asinh with Randomized Negative Value", "Biexponential Transformation",
    "Box-Cox Transformation", "FlowVS Transformation", "Hyperlog Transformation",
    "Linear Transformation", "Ln Transformation", "Log Transformation",
    "Logicle Transformation", "Quadratic Transformation", "Split Scale Transformation",
    "Truncate Transformation", "Centered Log Ratio Transformation", "None"),
  normalizationM = c("Bead-based Normalization", "GaussNorm", "WarpSet", "ZScore",
    "Mean Normalization", "Min-max Normalization", "None"),
  signalcleanM = c("FlowAI", "FlowCut", "PeacoQC", "None"),
  workflow = NULL,
  single_pos_fcs = NULL,
  single_pos_mass = NULL,
  beads_mass = c(140, 151, 153, 165, 175),
  sce_bead = NULL,
  marker_to_barc = NULL,
  min_cells = 3,
  max_bins = 10,
  step = 10,
  excludedColumn = NULL,
  DEP = NULL,
  clusteringM = "FlowSOM",
  ncluster = 8,
  Phenograph_k = 30,
  cores = parallel::detectCores()/2,
  save_processed_res = "one_folder",
  savepath = "./"
)
}
\arguments{
\item{name}{Character, the filename of the RData file when the "save_processed_res" parameter is set to "one_RData", and the filename of all the files in the "assess_res" folder which will store the assessment results.}

\item{datapath}{Character, the absolute path of the folder storing the FCS raw data files and metadata file.}

\item{mergeM}{Character, the method of merging multiple FCS files. When multiple FCS files are selected, cells can be combined using one of the four different methods including "Fixed", "Ceil", "All" and "Min".
\if{html}{\out{<br>}}\strong{Fixed}: a fixed num (specified by fixedNum) of cells are sampled (with replacement when the total number of cell is less than fixedNum) from each FCS file and combined for analysis.
\if{html}{\out{<br>}}\strong{Ceil}: up to a fixed number (specified by fixedNum) of cells are sampled without replacement from each FCS file and combined for analysis.
\if{html}{\out{<br>}}\strong{All}: all cells from each FCS file are combined for analysis.
\if{html}{\out{<br>}}\strong{Min}: The minimum number of cells among all the selected FCS files are sampled from each FCS file and combined for analysis.}

\item{fixedNum}{Integer, the fixed number of cells to be extracted from each FCS file.}

\item{compensationM}{Character, the method(s) of compensation for flow cytometry data including "CATALYST", "CytoSpill", "spillR" and "None". Compensation refers to the processing step of removing unwanted spillover resulting from signal crosstalk and spectral overlap across detection channels.
\if{html}{\out{<br>}}For details of each method, please refer to the \strong{Methods Introduction}.}

\item{transformationM}{Character, the method(s) of transformation for flow cytometry data including "Arcsinh Transformation", "Asinh with Non-negative Value", "Asinh with Randomized Negative Value", "Biexponential Transformation", "Box-Cox Transformation", "Centered Log Ratio Transformation", "FlowVS Transformation", "Hyperlog Transformation", "Linear Transformation", "Ln Transformation", "Log Transformation", "Logicle Transformation", "Quadratic Transformation", "Split Scale Transformation", "Truncate Transformation" and "None". Transformation refers to the processing step of adjusting the data with a heavily skewed distribution to a normal distribution.
\if{html}{\out{<br>}}For details of each method, please refer to the \strong{Methods Introduction}.}

\item{normalizationM}{Character, the method(s) of normalization for flow cytometry data, including "Bead-based Normalization", "GaussNorm", "Mean Normalization", "Min-max Normalization", "WarpSet", "ZScore" and "None". Normalization refers to the processing step of eliminating signal decay and technical variability across all files and batches over long-term data acquisition.
\if{html}{\out{<br>}}For details of each method, please refer to the \strong{Methods Introduction}.}

\item{signalcleanM}{Character, the method(s) of signal clean for flow cytometry data, including "FlowAI", "FlowCut", "PeacoQC" and "None". Signal cleaning refers to the processing step of identifying and removing abrupt signal shifts and changes that derive from (i) abrupt changes in the flow rate, (ii) clogs within the capillary tubes, (iii) temporary disruptions in cytometer fluidics, and (iv) unstable data acquisition.
\if{html}{\out{<br>}}For details of each method, please refer to the \strong{Methods Introduction}.}

\item{workflow}{Character, the combinations of data processing methods specified by users according to their research interests.
\if{html}{\out{<br>}}It is a vector includes one or more method combinations, typically in the format of "compensation method name_ transformation method name_ normalization method name_ signal clean method name ", for example: c("None_Biexponential Transformation_None_None","CytoSpill_FlowVS Transformation_None_FlowCut").}

\item{single_pos_fcs}{Character, the absolute filepath of the FCS file containing stained samples and control antibody-capture beads/pooled single-stained beads.
\if{html}{\out{<br>}}Only needed when "CATALYST" is included in the argument of "compensationM".}

\item{single_pos_mass}{Integer, the masses corresponding to barcode channels.
\if{html}{\out{<br>}}Only needed when "CATALYST" is included in the argument of "compensationM".}

\item{beads_mass}{Integer, the masses of the corresponding calibration beads.
\if{html}{\out{<br>}}Only needed when "Bead-based Normalization" is included in the argument of "normalizationM".}

\item{sce_bead}{SingleCellExperiment, the SingleCellExperiment object for the bead experiment.
\if{html}{\out{<br>}}Only needed when "spillR" is included in the argument of "compensationM".}

\item{marker_to_barc}{Data frame, the table that maps the marker to the barcode in the beads experiment.
\if{html}{\out{<br>}}Only needed when "spillR" is included in the argument of "compensationM".}

\item{min_cells}{Integer, the minimum amount of cells (nonzero values) that should be present in one bin.
\if{html}{\out{<br>}}Only needed when "PeacoQC" is included in the argument of "signalcleanM". Lowering this parameter can affect the robustness of the peak detection.}

\item{max_bins}{Integer, the maximum number of bins that can be used in the cleaning process.
\if{html}{\out{<br>}}Only needed when "PeacoQC" is included in the argument of "signalcleanM". If this value is lowered, larger bins will be made.}

\item{step}{Integer, the step in events_per_bin to which the parameter is reduced to.
\if{html}{\out{<br>}}Only needed when "PeacoQC" is included in the argument of "signalcleanM".}

\item{excludedColumn}{Character, the non-protein columns names of which accessed through the function "Getmarker".
\if{html}{\out{<br>}}It is a string separated by commas, typically in the format of "channel description (channel name)", for example: "gate_source(gate_source), cell_id(cell_id), sample_id(sample_id), Time(Time), Cell_length(Cell_length), DNA-1(DNA.1.Ir191.Dd)".}

\item{DEP}{Character, the absolute filepath of the CSV file including the differentially expressed proteins used as the prior knowledge for the fourth criterion.
\if{html}{\out{<br>}}It is a table of one column without the column name, each table cell includes one protein typically in the format of "channel description (channel name)", for example: "CD20(FITC.A)".}

\item{clusteringM}{Character, the method of clustering the processed data prior to plotting, including "FlowSOM" and "PhenoGraph".}

\item{ncluster}{Integer, the number of clusters for meta clustering in "FlowSOM".
\if{html}{\out{<br>}}Only needed when the argument of "clusteringM" is selected as "FlowSOM" while calling the function "FCprocess" or "MCprocess" for obtaining "data".}

\item{Phenograph_k}{Integer, the number of nearest neighbours used in PhenoGraph clustering method.
\if{html}{\out{<br>}}Only needed when the argument of "clusteringM" is selected as "PhenoGraph".}

\item{cores}{Integer, the number of CPU cores to be employed for performing parallel computing.
\if{html}{\out{<br>}}To avoid memory explosion due to parallel computing, the default is the largest integers not greater than half of the number of CPU cores on the current host.}

\item{save_processed_res}{Character, the format of the data processing output files. "no" denotes that the results would not be saved. "one_folder" denotes that successfully processed results will be saved as separate RData files in the "process_res" folder. "one_RData" denotes that all processed results will be saved as one RData file in the "process_res" folder.}

\item{savepath}{Character, the absolute path of the folder which will store files of the assessment and processed results.}
}
\value{
The data processing and performance assessment output files, which are located in the \strong{process_res} and \strong{assess_res} folders, respectively.
\subsection{Results of Data Processing}{

The \strong{process_res} folder stores the results of various data processing workflows. The form of data processing output files is decided by the parameter \code{save_processed_res}: "no" denotes that the results would not be saved; "one_folder" denotes that successfully processed results will be saved as separate RData files in the "process_res" folder; "one_RData" denotes that all processed results will be saved as one RData file in the "process_res" folder [\emph{default ="one_folder"}].
}

\subsection{Results of Performance Assessment}{

The \strong{assess_res} folder stores the results of performance assessment, which includes 3 files:
\if{html}{\out{<br>}}(\strong{1}) \strong{_Ranking_Table.csv} contains the overall ranking results of all data processing workflows, where the "Rank" column represents the overall ranking, and the "Value" column shows the scores for different assessment criteria.
\if{html}{\out{<br>}}(\strong{2}) \strong{_Ranking_Figure.pdf} visualizes the overall ranking results to help users better understand the differences among various data processing workflows. Different colors represent different performance assessment levels: dark green indicates "superior," light green indicates "good," and red indicates "poor."
\if{html}{\out{<br>}}(\strong{3}) \strong{_assess.RData} contains 2 lists, "table" and "table2", providing the raw scores for different assessment criteria and performance assessment levels categorized by thresholds, respectively.
}

\subsection{Records}{

In addition, the files \strong{log.txt} and \strong{info_saved.RData} are also generated simultaneously. \strong{log.txt} records the processing details while \strong{info_saved.RData} records the information related to "metadata" and "index_protein".
}
}
\description{
MC_CSI enables high-throughput processing of SCP data acquired from MC using up to ~675 available workflows and subsequently assesses the performance of all processing workflows based on comprehensive criteria from the perspective of CSI studies, functioning similarly to a combination of MCprocess() and CSIassess(), while utilizing less memory during execution.
}
\examples{
\donttest{
}
}
