% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{Process}
\alias{Process}
\title{Process}
\usage{
Process(
  name,
  datapath,
  techique = c("MC", "FC"),
  studytype = c("CSI", "PTI"),
  mergeM = c("Fixed", "Ceil", "All", "Min"),
  fixedNum = 200,
  compensationM = c("AutoSpill", "CATALYST", "CytoSpill", "FlowCore", "MetaCyto", "None"),
  transformationM = c("Arcsinh Transformation", "Asinh with Non-negative Value",
    "Asinh with Randomized Negative Value", "Biexponential Transformation",
    "Box-Cox Transformation", "FlowVS Transformation", "Hyperlog Transformation",
    "Linear Transformation", "Ln Transformation", "Log Transformation",
    "Logicle Transformation", "Quadratic Transformation", "Scale Transformation",
    "Truncate Transformation", "None"),
  normalizationM = c("Bead-based Normalization", "GaussNorm", "WarpSet", "ZScore",
    "None"),
  signalcleanM = c("FlowAI", "FlowClean", "FlowCut", "None"),
  spillpath = NULL,
  FSC = "FSC-H",
  SSC = "SSC-H",
  control.dir = NULL,
  control.def.file = NULL,
  single_pos_fcs = NULL,
  single_pos_mass = NULL,
  CATALYSTM = c("flow", "nnls"),
  beads_mass = c(140, 151, 153, 165, 175),
  index_protein = NULL,
  save_processed_res = c("no", "one_folder", "one_RData"),
  savepath = "./",
  cores = floor(parallel::detectCores()/2)
)
}
\arguments{
\item{cores}{}
}
\description{
Process
}
\examples{
\donttest{
}
}
