% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTIassess.R
\name{PTIassess}
\alias{PTIassess}
\title{PTI (Pseudotime Trajectory Inference) Assess}
\usage{
PTIassess(
  data,
  respath,
  TIM = c("scorpius_distSpear", "scorpius_distPear", "scorpius_distEucl",
    "scorpius_distManh", "slingshot_FLOWMAP", "slingshot_tSNE", "prinCurves_tSNE",
    "slingshot_PCA", "slingshot_diffMaps", "prinCurves_diffMaps"),
  Cc_metric = c("Spearman rank correlation", "Kendall rank correlation"),
  pathwayhierarchy = NULL,
  save_processed_res = c("no", "one_folder", "one_RData"),
  cores = floor(parallel::detectCores()/2),
  ...
)
}
\arguments{
\item{data}{Character, the resulting RData file of the function "Process", "FCprocess" or "MCprocess" when the "save_processed_res" parameter in these functions is set to "one_RData".}

\item{respath}{Character, the absolute path of the folder storing the resulting RData files of the function "Process", "FCprocess" or "MCprocess" when the "save_processed_res" parameter in these functions is set to "one_folder".}

\item{TIM}{The method of trajectory inference for the processed data prior to performance assessment, consisted of trajectory reconstruction and data space representation, including "scorpius_distSpear", "scorpius_distPear", "scorpius_distManh","slingshot_tSNE", "prinCurves_tSNE", "slingshot_PCA", "slingshot_diffMaps" and "prinCurves_diffMaps".}

\item{Cc_metric}{Character, the assessing metric under Criterion Cc for the "PTI" study type, including "Spearman correlation" and "Kendall Rank Correlation".
\if{html}{\out{<br>}}\strong{Spearman correlation}: a metric that measures the monotonic relationship between two ranked variables by measuring how well the relationship between the variables can be described by a monotonic function, with values closer to ±1 indicating a stronger monotonic association.*
\if{html}{\out{<br>}}\strong{Kendall Rank Correlation}: a metric that evaluates the similarity between two rankings by calculating the proportion of concordant and discordant pairs, with higher values indicating a stronger agreement in the relative ordering between the variables.}

\item{pathwayhierarchy}{Character, the absolute filepath of the pathway hierarchy file.}

\item{save_processed_res}{Character, the form of data processing output files. "no" denotes that the results would not be saved. "one_folder" denotes that successfully processed results will be saved as separate RData files in the "process_res" folder. "one_RData" denotes that all processed results will be saved as one RData file in the "process_res" folder.}

\item{cores}{Integer, the number of CPU cores to be employed for performing parallel computing.
\if{html}{\out{<br>}}To avoid memory explosion due to parallel computing, the default is the largest integers not greater than half of the number of CPU cores on the current host.}

\item{name}{Character, the filename of all the files in the "assess_res" folder which will store the assessment results.}

\item{savepath}{Character, the absolute path of the folder which will store files of the assessment and processed results.}
}
\value{
A list containing 2 tables, "table" and "table2", which provide the raw scores for different assessment criteria and performance assessment levels categorized by thresholds, respectively.
}
\description{
PTIassess assesses processing performance of all workflows which are used while running the function "FCprocess" or "MCprocess" based on comprehensive criteria (each with distinct underlying theories) from the perspective of PTI studies.
}
\examples{
\donttest{
}
}
