% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{Visualize}
\alias{Visualize}
\title{Visualize}
\usage{
Visualize(
  studytype = c("CSI", "PTI"),
  respath,
  save_processed_res = "one_folder",
  workflow,
  savepath = "./ANPELA_res",
  studyname = NULL,
  plot_metric = c("Ca_metric", "Cb_metric", "Cc_metric", "Cd_metric"),
  clusteringM = c("FlowSOM"),
  ncluster = 8,
  ntop = NULL,
  DEP = NULL,
  TIM = "scorpius_distSpear",
  clustering.var = NULL,
  pathwayhierarchy = NULL,
  Cc_metric = "Spearman rank correlation"
)
}
\arguments{
\item{studytype}{Character, the type of your study, including ‘CSI (Cell Subpopulation Identification)’ and ‘PTI (Pseudotime Trajectory Inference)’.}

\item{respath}{Character, the absolute path of the folder storing the resulting ‘info_saved.RData’ file and the ‘process_res’ folder of the function ‘Process’, ‘FCprocess’ or ‘MCprocess’ when the ‘save_processed_res’ parameter in these functions is set to ‘one_folder’.}

\item{save_processed_res}{Character, the format of the data processing output files. ‘one_folder’ denotes that successfully processed results will be saved as separate RData files in the ‘process_res’ folder. ‘one_RData’ denotes that all processed results will be saved as one RData file in the ‘process_res’ folder.}

\item{workflow}{Character, the combinations of data processing methods specified by users according to their research interests.}

\item{savepath}{Character, the absolute path of the folder which will store the results.}

\item{studyname}{Character, the filename of the PNG files generated that includes the value of this required character string parameter, serving as a user-specified identifier for the analysis run.}

\item{plot_metric}{Character, the metric(s) which will be visualized and saved.}

\item{clusteringM}{Character, the method of clustering the processed data prior to performance assessment, including ‘FlowSOM’ and ‘PhenoGraph’.}

\item{ncluster}{Integer, the number of clusters for meta clustering in FlowSOM.}

\item{ntop}{Integer, the number of the most differentially expressed markers that are truncated for calculating the CWrel value.}

\item{DEP}{Character, the absolute file path of the CSV file including the differentially expressed proteins used as the prior knowledge for the fourth criterion.
\if{html}{\out{<br>}}It is a table of one column without the column name, each table cell includes one protein typically in the format of "channel description (channel name)", for example: "CD20(FITC.A)".}

\item{TIM}{Character, the method of trajectory inference for the processed data prior to performance assessment, consisted of tra-jectory reconstruction and data space representation, including ‘scorpius_distSpear’, ‘scorpius_distPear’, ‘scorpius_distEucl’, ‘scorpius_distManh’, ‘slingshot_tSNE’, ‘slingshot_FLOWMAP’, ‘slingshot_PCA’, ‘slingshot_diffMaps’.}

\item{clustering.var}{Character, the vector naming channels to be used to calculate distances/differences between cells for clustering (if re-quested) and edge-drawing steps.}

\item{pathwayhierarchy}{Character, the absolute file path of the pathway hierarchy file.}

\item{Cc_metric}{Character, the assessing metric under Criterion Cc for the ‘PTI’ study type, including ‘Spearman correlation’ and ‘Kendall Rank Correlation’.}
}
\value{
PNG files visualizing the assessment standards.
}
\description{
Visualize() loads processed SCP data and generates a suite of visualizations designed to evaluate the quality and characteristics of either CSI or PTI results, sav-ing these plots for review.
}
\examples{
\donttest{
}
}
