% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTIassess.R
\name{PTIassess}
\alias{PTIassess}
\title{PTI (Pseudotime Trajectory Inference) Assess}
\usage{
PTIassess(
  name = "result",
  data,
  respath,
  TIM = c("scorpius_distSpear", "scorpius_distPear", "scorpius_distEucl",
    "scorpius_distManh", "slingshot_tSNE", "slingshot_FLOWMAP", "slingshot_PCA",
    "slingshot_diffMaps"),
  Cc_metric = c("Spearman rank correlation", "Kendall rank correlation"),
  pathwayhierarchy = NULL,
  save_processed_res = "one_folder",
  savepath = "./ANPELA_res",
  clustering.var = NULL,
  cores = floor(parallel::detectCores()/2),
  ...
)
}
\arguments{
\item{name}{Character, the filename of the RData file in the "assess_res" folder which will store the assessment results.}

\item{data}{Character, the R object resulting from the function "Process", "FCprocess" or "MCprocess", or obtained by loading from the resulting RData file of these funcitons when the \code{save_processed_res} parameter in these functions is set to "one_RData".}

\item{respath}{Character, the absolute path of the folder storing the resulting "info_saved.RData" file and the "process_res" folder of the function "Process", "FCprocess" or "MCprocess" when the \code{save_processed_res} parameter in these functions is set to "one_folder".}

\item{TIM}{Character, the method of trajectory inference for the processed data prior to performance assessment, consisted of tra-jectory reconstruction and data space representation, including ‘scorpius_distSpear’, ‘scorpius_distPear’, ‘scorpius_distEucl’, ‘scorpius_distManh’, ‘slingshot_tSNE’, ‘slingshot_FLOWMAP’, ‘slingshot_PCA’, ‘slingshot_diffMaps’.}

\item{Cc_metric}{Character, the assessing metric under Criterion Cc for the "PTI" study type, including "Spearman correlation" and "Kendall Rank Correlation".
\if{html}{\out{<br>}}\strong{Spearman correlation}: a metric that measures the monotonic relationship between two ranked variables by measuring how well the relationship between the variables can be described by a monotonic function, with values closer to ±1 indicating a stronger monotonic association.*
\if{html}{\out{<br>}}\strong{Kendall Rank Correlation}: a metric that evaluates the similarity between two rankings by calculating the proportion of concordant and discordant pairs, with higher values indicating a stronger agreement in the relative ordering between the variables.}

\item{pathwayhierarchy}{Character, the absolute filepath of the pathway hierarchy file.}

\item{save_processed_res}{Character, the format of the data processing output files. "no" denotes that the results would not be saved. "one_folder" denotes that successfully processed results will be saved as separate RData files in the "process_res" folder. "one_RData" denotes that all processed results will be saved as one RData file in the "process_res" folder.}

\item{savepath}{Character, the absolute path of the folder which will store the assessment results.}

\item{clustering.var}{Character, the vector naming channels to be used to calculate distances/differences between cells for clustering (if re-quested) and edge-drawing steps.
\if{html}{\out{<br>}}Only needed when ‘slingshot_FLOWMAP’ is included in the parameter of ‘TIM’.}

\item{cores}{Integer, the number of CPU cores to be employed for performing parallel computing.
\if{html}{\out{<br>}}To avoid memory explosion due to parallel computing, the default is the largest integers not greater than half of the number of CPU cores on the current host.}
}
\value{
The \strong{assess_res} folder stores the assessment output file named \code{name}\strong{_assess.RData}, which contains 2 lists, "table" and "table2", providing the raw scores for different assessment criteria and performance assessment levels categorized by thresholds, respectively.
\if{html}{\out{<br>}}In addition, the file \strong{log.txt} is also generated simultaneously, recording the processing details.
}
\description{
PTIassess() assesses processing performance of all workflows which are used while running the function "FCprocess" or "MCprocess" based on comprehensive criteria (each with distinct underlying theories) from the perspective of PTI studies.
}
\examples{
\donttest{
}
}
